packages/llm-service/pyproject.toml
[tool.poetry]
name = "llm-service"
version = "0.1.0"
description = ""
authors = ["Rahel Lüthy <rahel.luethy@fhnw.ch>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "<=0.24.0"
langserve = {extras = ["server"], version = ">=0.1.0"}
pydantic = "<2"
langchain = "^0.1"
tiktoken = ">=0.5.1"
chromadb = "==0.5.3" # 0.5.4 breaks the app -> https://github.com/langchain-ai/langchain/issues/24163
beautifulsoup4 = ">=4.12.2"
langchain-openai = "^0.1.3"
lark = "^1.1.9"
scikit-learn = "^1.4.1.post1"
pandas = "^2.2.1"
jinja2 = "^3.1.3"
matplotlib = "^3.8.3"
umap-learn = "^0.5.6"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.21"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
